fix: request #43344 Artifact link field check update permission even if no changes
The new artifact link field always send a value even it there is no
changes. As the artifact view form works like a PATCH request it
consider that if there is data for a field the user submitted changes to
it and thus we check submit/update permission.

In the case the field have no changes but user doesn't have the
permission to update it, the permission is still checked and the
validation fail.

*Testing:*

On a tracker with the link field, set the submit permission to project
members and the update permission to project admins.

With a project member update an artifact without editing the link field
-> the request should pass.

Change-Id: I3d044cc9ae59525c84ba4fa5ab0f5a68396bac04