refactor: replace store by collection for allowed trackers
part of story #42248 Artidoc read-only fields: User Groups list fields

No functional change expected. You can still  configure an artidoc and
choose its tracker from a list (filtered by the backend). When there are
no trackers in the project (or there is one but it has one required
field that is neither Title nor Description), both the configuration
Modal and Panel still show an error (with disabled submit) to explain
that there are no trackers.

Why?
Replaced the equals check with a dedicated method.
I chose to move the Tracker type to the collection of allowed trackers,
but it could have been standalone.
It would be nice to remove the type guard on
isTrackerWithSubmittableSection, but TypeScript only narrows the type
when doing the checks inline, not wrapped in a function. It might be
better with a more recent version of TypeScript, but that's another
story.

Change-Id: Idc7bf5eef385bc69ae87d9586c9adb4e036d40c8