chore: Migrate eslint to flat config
part of request #42992 eslint: 8.52.0 -> 9.x

How to test:
No functional change. CI should be happy. Configurations should still
flag code that breaks the rules in:
- plain javascript files
- typescript files
- Vue files (both with JS / TS)
- storybook stories
- cypress tests
- JS tests
- TS tests
- Node files (webpack, Vite, Jest configurations)

Why?
eslint config is now .mjs to force using ES module imports. So is the
config for stylelint.

I chose to remove the "you don't need lodash" plugin. Lodash uses are
all in this review, we expect to have to change this code for the
migration out of AngularJS, it will be an opportunity to replace them by
native functions.

I disabled "cypress/unsafe-to-chain-command": it flags several errors in
our cypress tests, and it has been long enough to do this migration. We
will fix them in a dedicated request.

@typescript-eslint packages are replaced by a single, not namespaced
"typescript-eslint" package.

Change-Id: I33fd1130fbee3af9e29b8428728750175bfd6c66