refactor: Nuke Tracker_Semantic_Description::load()
part of story #42248 Artidoc read-only fields: User Groups list fields

Why?

99% of the time, the semantic is retrieved to know whether it is defined
or what's the field bound to it, and hardly to get benefits of the whole
TrackerSemanticDescription object.

Moreover, the fact it can be retrieved using a static method is not
great because it forces us to have setInstance/clearInstances to mock it
in unit tests, making them harder to write and maintain.

Instead, let's have a cache object (as a singleton) coupled to a
description field retriever to achieve the same thing without having to
load the whole semantic object.

-- no functional change expected

Change-Id: I11383e663fede1f21a3851923c54c255616c3070