refactor: Extract is_{error,success,saving} from configuration store
Part of story #42246 Artidoc read-only fields: Static List fields

`saveTrackerConfiguration` and `saveFieldsConfiguration` now return a
ResultAsync. This way configuration component can use it to update their
refs.

In the case of ConfigurationModal it was using is_success. So I choose
to use a computed ref which get the value in child components

*No functional changes expected*

Change-Id: Iae8865c1d8b28458df4b0d091e894bbc00b0342e