refactor: widget_id is misleading
When we are doing db queries we are refering to widget_id. But what
widget_id is referring to? The id of the widget, or the id of the query?

plugin_timetracking_management_query(id, ...)
plugin_timetracking_management_query_users(widget_id, ...)

query_users is needed because we need to store many users for a
configuration. But query_users.widget_id is referring to query.id...
Whereas one might think that it refers to dashboards_lines_columns_widgets.id.

In order to remove the ambiguity, widget_id is now query_id.

No functional changes

Part of story #41084: Retrieve widget configuration

Change-Id: Ie2ecf5f55fb045efe69efa17c41dae72acb9a3d8