feat: Close the configuration modal after fields save
Part of story #42246 Artidoc read-only fields: Static List fields

After saving readonly fields configuration, the modal should close as
there is nothing more to do (+ it allows to see result). In case of
error, the feedback is still displayed and the modal not closed

The tracker tab should not be impacted

Also moved all save logic inside footers

*Testing:*

In an artidoc with a tracker configured, do some changes to the readonly
fiels and save -> the modal should close.

Change-Id: I96b5a9ad13d32eb5323f46a151080dc0a0afea89