refactor: remove notion of UserDiff
When the client of the api is sending a list of users for a query, there
is no need to know which users have been removed and which ones have been
added. We just have to clean the previous setting and save the new list.

No functional changes

Part of story #41084: Retrieve widget configuration

Change-Id: I8c36677080b904c323b2a5da31c52ca20afee727