refactor: Nuke TrackerSemanticTitle::load()
Part of story #42247 Artidoc read-only fields: Users list fields

99% of the time, the semantic is retrieved to know whether it is defined
or what's the field bound to it, and hardly to get benefits of the whole
TrackerSemanticTitle object.

Moreover, the fact it can be retrieved using a static method is not
great because it forces us to have setInstance/clearInstances to mock it
in unit tests, making them harder to write and maintain.

Instead, let's have a cache object (as a singleton) coupled to a
title field retriever to achieve the same thing without having to
load the whole semantic object.

*No functional changes expected*

Change-Id: If0954dbaa2823248a98a0fb638256aec3b4841f1