fix: Restricted users should be viewable
Managers should be able to see times for active and restricted users.

Took the opportunity to fix Restler anotations so that we can have nice
@array docblocks.

Unit tests are now added to have a better safety net on this sensitive
part.

This contribution started with the introduction of current user ability
to view user, so a basic test with anonymous check has been added, it
will be complemented in next contributions.

Part of story #41084: Retrieve widget configuration

Change-Id: Ib2a3a0dbbaad1527211bacc0cd80aeada1659263