chore: request #43697 Move base image from Rocky Linux 9.4 to 9.6
No real changes since we make sure to updates the base packages
but we might as well use the current image.

Also took the opportunity to remove the single layer trick.
It is not really useful anymore as we are depending since
the base image is far leaner (and recent) than when we introduced it.

You can locally build the TCE image with something like:
docker build -f "$(pwd)/tools/docker/tuleap-community-edition/Dockerfile" "$(pwd)/tools/"

Change-Id: I4ac5a94d27f6bd17a096fda7aa3eb7a912863ed3