feat: Display link field
Part of story #42249 Artidoc read-only fields: artifact links

How to test:
- Forge the database table plugin_artidoc_document_tracker_field and add
a Links field for the configured tracker of your artidoc.
- Add some artifact sections to your artidoc.
- Set up artifact links between the artifacts.
- You should see the links rendered as a "block" (full width of the
page). Each link shows the title (if any), the cross-reference (tracker
shortname, artifact ID, tracker color) and the status badge (label +
color, if any).

Why?
We had to change the type of the display type "enum" so that Artifact
links fields are always "block".

Co-authored-by: Joris Masson <joris.masson@enalean.com>
Co-authored-by: Thomas Gorka <thomas.gorka@enalean.com>

Change-Id: I8e0376bce759a765d2c5947dc112aecd094d9c41