feat: remove users that can no longer be seen 🛂
The users that can be selected in the widget must be restricted to users
that have TIMETRACKING WRITE permissions in the projects where the
manager has TIMETRACKING READ permission. The selector is not changed in
this contribution: only the backend is changed on save of configuration.

Given two users in the selection from a previous configuration: `timtra`
and `deschcla`
When the manager update the list to add `lucdem` and the manager has no
longer the access rights to see `timtra` times
Then, on submit, `timtra` is removed from the selection and a warning is
displayed to inform the manager.

⚠️ This contribution is dealing with sensitive information (time spent by
user is a PII) and cross boundaries aggregation (projects, trackers, …).
It needs special care in the review to ensure that every scenario has
been handled. Suggested entry point: integration test
ManagerCanSeeTimetrackingOfUserVerifierDaoTest.

Part of story #41084: Retrieve widget configuration

Change-Id: I9b13064cc122334014a28ee823e5e43fe74ae297