refactor: Use Option for Domain objects
Part of story #42249 Artidoc read-only fields: artifact links

No functional change expected. When you forge your DB to configure an
Artifact Links field for an Artidoc, you should still retrieve the links
with their status (if any) and their status color (if any).

Why?
REST representations are constrained to using null because Option does
not exist in JSON. However, our Domain objects should use Option to
better communicate and represent values that can be present or absent.

I am forced to use psalm-suppress in the representation so that we can
keep it @psalm-immutable, because I cannot make Option::mapOr()
mutation-free (it could call some DAO or whatever and cause
side-effects). We can see however that in this instance it is really
mutation-free: in one case, it returns the name of the ItemColor, in the
other it turns a domain object into a representation.

Change-Id: Ic71ca5a0ee8c0663637b242c902cbb5cd478785d