refactor: Build planning admin colorpicker separately
Part of request #30339 Split agiledashboard frontend into separate apps

How to test:
No functional change expected in the Backlog service administration,
when you edit a Planning. When Cardwall or Taskboard is enabled, it
should load correctly the color picker to assign a color to a Cardwall
column.
No functional change on the color picker in Tracker fields admin, when
editing a static List field.

Why?
It should be Cardwall plugin that loads this script, but I'm not here
for this mess. It's better than before because it is built by Vite, and
it's a separate root file, and it's now loaded only on the "Edit"
planning page, since when you create a new planning, you cannot do
anything with regards to Cardwall columns.
I removed some CSS classnames that were no longer used, they belonged to
the "old" color picker with RGB colors.

Change-Id: I9e231f71e25738f12098392bc2b039540023c013