refactor: Build FP styles for Backlog separately
Implements request #30339 Split agiledashboard frontend into separate
apps

How to test:
No functional change expected in the Backlog service administration:
when you create a new planning, edit a planning, or configure the burnup
chart, the FlamingParrot styles are still loaded.
end-to-end tests can still be typechecked.

Why?
I could not find usage of "unlink.png", and it is not publicly reachable
(not in "www"), therefore it is assumed dead.
The "not configured" images were used only on the AgileDashboard
homepage (that has been dead for quite a few months). Same for the
"tuleap-artifact-modal" rules, it was for the Modal v2 that was only on
the AgileDashboard homepage.
Loading the styles on the correct pages should be done separately, it
still depends on "cssfile" event

Change-Id: Ia1a87a8e64d672f1aff971f4f8b47f854c3f43b4