feat: Display project of linked artifact
Part of story #42249 Artidoc read-only fields: artifact links

How to test:
- Configure your artidoc using the REST route for configuration, add
an artifact links field for the configured tracker of your artidoc.
- Add some artifact sections to your artidoc.
- Link an artifact from the same project and an artifact from another
project to the section.
- For the link in the other project, you should see the project label
and icon appear under the artifact title. If there is no icon, there
should be no empty space before the label.
- For the link in the same project, the project label is not shown (it
is implicitly the current project).

Why?
I chose not to reuse \Tuleap\Project\REST\ProjectReference because we
don't need the URI to the project and because the representation builder
would need to retrieve again the Project object (which is needed to
build it). I cannot put a Project in ArtifactLinkValue, because it's
outside the Domain.
If the URI becomes needed, we should trash ArtifactLinkProject and
replace it by ProjectReference.

Change-Id: Ic0bbb3fb343499c0baf29f9f0c11932e57b194ca