fix: request #44016 Disable timeout when starting Tuleap systemd service
On deployments hosting Tuleap and the DB on the same server,
the DB can take some time to become accessible after a reboot.
In some situations we can reach the timeout of the service we will
then require an human intervention to start.
Raising the timeout should not be an issue, it still quite easy to
determine that something is wrong if the service cannot start because
Tuleap will not be accessible.

To test the change on your dev instance the easiest way is to directly
edit `/usr/lib/systemd/system/tuleap.service` and then
`systemctl daemon-reload && systemctl restart tuleap`.

Change-Id: I7ddf1e1538ccb7c8f5c82d862e529312b483d9e7