chore: Prevent AngularJS to rely on eval()
No functional changes expected, to test it works as expected you
can remove the `unsafe-eval` directive from the CSP in the `pre.php`.

This might have some performance impact but nothing that should be
significative for our use cases. If it is the proper fix is to remove
AngularJS not to let it do awful things with `eval()`.

Part of request #17967: Deploy a useful content security policy

Change-Id: Id3c0e8b0670a9bc656b2a8734e48e8467938fc96