fix: request #44023: New file cut off in PR unified diff
When reviewing PR the complete file for an added file is not shown in
unified diff. In side by side diff the whole file is shown.

When the bug occurs, just clicking in the empty area, or resizing the
viewport force the redraw of the editor and missing lines are displayed.
This contribution force the redraw by moving the cursor at the end of
the editor and then moving it back to the beginning.

Change-Id: I15ad8f9771a72df15175930c26b1294789050d86