refactor: adapt TaskBar tests
Part of request #41441 Migrate Roadmap widget to Vue 3

The tests in TestBar checked the styles of the TaskBarProgress component. The problem is that in Vue 3 we no longer have access to the HTML of child components, as the current component is stubbed.
In this patch I moved the tests to TaskBarProgress.test.ts so that they could check the required styles.

No functional changes expected

Change-Id: Ibe7feddc7079096203a1fbefa32a4740badd7a12