refactor: Nuke TrackerSemanticStatus::load
Part of story #44017 Artidoc read-only fields: Numeric fields

The usage of static method to retrieve important object like this one is
not great as it forces us to use setInstance/clearInstances for testing.

In general it is better to have instantiated class we can inject through
their interfaces and then stub in tests.

*No functional changes expected*

Change-Id: I2c2bcdd28add80669a771248b7fdc6c24a757f35