fix request #44030 Burnup data built and cached data cannot be the same
Cached data of the start date is stored with the time 11:59 pm (23:59) while in the data builder the time is 00:00 am (00:00).

The consequence is that the cache is invalid and all the values need to be calculated again.

For some reason we rebuilt a new start date at 00:00 (and not 23:59) and
use this new start date value to validate or not the cache.

Change-Id: Ia46b6ba8f56a83f6b3b88816ae0821813146cf5e