refactor: request #44046 Mark overriding methods with attributes
No functional change. CI should be happy.

Why?
Marking those methods should make it clear to readers that there is a
parent method somewhere whose behaviour is overridden. PHP will raise an
error if the parent method is deleted, this should provide a warning to
developers that there's something more to investigate.
Existing calls where an attribute would be required are added to the
Psalm baseline. We should fix them gradually (we can also fix them in
bulk, but it's more risky).

Change-Id: I1babc7e04b0e8374df0e8b1000c68480fbd61fda