refactor: kill RetrieveSemanticStatusFieldIterativeStub
Part of story #44017 Artidoc read-only fields: Numeric fields

How to test: No functional change, CI should be happy.

Why?
RetrieveSemanticStatusFieldStub has a more realistic behaviour: it can
be prepared with fields, and will return them if they match the given
tracker. Previously, to achieve the same, we used a "withSuccessiveXXX"
method, but it makes the test brittle. If the order of calls changes,
the test will break.

Since I was working on ArtifactCloserTest and
PostPushWebhookCloseArtifactHandlerTest, I took the opportunity to
change some stub calls. When we have stubs that hide side-effects (such
as saving something to the DB), we can prepare them with a callback.
Instead of telling the stub to give us their data, we "just" do the
assertions we need in the callback. Thus, CreateCommentOnlyChangesetStub
now functions almost exclusively with callbacks and we can remove its
getters. Same for CreateNewChangesetStub, we don't need a getter
anymore.

I also changed ListStaticBindBuilder so that when called with BindValue
objects, it deals with the array structure by itself. Callers can just
give it a list of built values, and it will arrange them on its own.

Change-Id: I0a3713bf87ad8d9c9aa7d5ca8da69d5f6b2cb769