fix: request #44057 Running tuleap-cfg site-deploy can fail depending on the current directory
If the current working directory is not accessible running the upgrade
processes needed by MediaWiki will fail.
Since PHP 8.3 `proc_open()` will be unhappy if the current working
directory does not exist/is not accessible, this can typically be the
case with automated upgrade processes.

Since we do not really care about the current working directory in the
situation where we call `proc_open()`, the current working directory is
now set to '/'. The change in behavior of PHP should only be an issue
for commands launched from CLI tools but it seemed easier to adjust
everything.

Change-Id: I52f72e476eafe5a173f04572b9bbc89101a2bb63