fix: tracker_changeset_value_artifactlink goes 3NF
When an artifact is being moved in another tracker with a different
name, then there are some places (e.g. links column in tracker report,
links changes in artifact follow-ups, …) where the reference to the
artifact has a misleading keyword (name of the previous tracker).

Scenario:

1. Have a `sprint` and a `task` tracker
2. Have a `task #123` linked to `sprint #101` (replace with your own ids)
3. Create a `copytask` tracker from `task` tracker
4. Move `task #123` to `copytask`
5. Every references displayed for artifact 123 should be `copytask #123`

This is due to the fact that the database schema contains duplication.
When we have an artifact id stored in the artifact link, there is no
need to store as well the project id or the keyword of the reference.
Those can be deducted by joining few tables (which is already the case
BTW). If an artifact is moved, we would have to update every
artifact links to reflect the new keyword and project id. Applying third
normal form (3NF) to tracker_changeset_value_artifactlink removes this
need and resolves the issue.

Closes request #44058: Misleading reference name after moving an artifact

Change-Id: I099033648518f3a8f145e3000f542aa88c4b0491