feat: add a load all button to display all artifact links
Part of story #44054 Cross-tracker search: load all button

The goal is to reduce the risk of overloading client and server if there are too many links attached to one artifact.
By default, the expand loads 50 forward links and 50 reverse links.
If there are more than 50 forward or reverse links a "load all" button is displayed.

Note: A confirmation modal will be added in the next patch.

How to test:
-- Have an XTS widget with at least an artifact with more than 50 forward links
-- Click on the caret of this artifact to display its links
=> It should display a maximum  of 50 forward links
=> It should display a 'load all' button at the end

-- Click on the 'load all' button
=> It should display all the forward links
=> It should not display the 'load all' button anymore

Change-Id: I7ac222070c425e14c479110ca1eb45ef38ec0a11