fix: Changeset value may be null
Part of story #44017 Artidoc read-only fields: Numeric fields

When retrieving changeset_value of a field, the returned value may be
null. This case is not handled for string and list field:

- For string field the case is skipped and the field is not displayed at
  all
- For list field we have a TypeError

Now it will just display Empty.

The case of a null changeset_value can happen on newly created artifact
from API -> non given field are not set

*Testing:*

The easiest way is to forge your db by removing some changeset_value

The corresponding field value should be displayed as Empty

Change-Id: Ic9d063f932b7d04481898c0d06e1efba1ed37ba0