feat: log invalid user ids
Add a user to the timetracking management widget. Save the query.
=> the user is still here

As siteadmin, switch the user status to suspended.

Edit the timetracking management widget, do not change anything, save
the query.
=> the user disappears from the list

Look at rest api logs, there should be a mention that the user does not
exist or is not active nor restricted.

Part of story #41084: Retrieve widget configuration

Change-Id: I3423345af9a8fa142bc7718a4d0f68dddd699546