fix: It should display the correct number of skeletons
Part of story #44054 Cross-tracker search: load all button

What has been fixed?
The number of skeletons displayed corresponded to the sum of the expected number of forward links + the expected number of reverse links.
Since the last patch, we display the maximum (50) number of forward and reverse links, and if there are more, a “load all” button is displayed.
To avoid sudden changes in the display, the number of skeletons to be displayed is the minimum between the expected number of artifact links and 50.

Change-Id: Ied6f93388c186a3fa2ec1fa18af90aa2d6dd270c