refactor: Make description stub more realistic
Part of story #44019 Artidoc read-only fields: User fields

How to test: No functional change. CI should be happy.

Why?
Recently we introduced a stub for Title semantic that can be setup with
several fields and that will respond based on its setup field's tracker
ID, instead of blindly always returning the same field. We should do the
same for description, because semantics are often used together.
Besides, it forces us to write more realistic tests: if we forget to say
which tracker a field belongs to, the code will fail to find the
description field for the tracker we will give.
I also took the liberty of moving the setup of some tests to a single
function with class properties. It makes it more convenient to find what
is really changing between each test cases, and we only need to override
the default setup once or twice anyway.

Change-Id: Iaee19e8df82f6e47fd0761379b4e14e0881e4f94