fix!: rename REST route to update widget
When we change the "query" of a timetracking management widget, we
actually change the configuration of the widget.

This introduce a breaking change in the API, but the feature is hidden
by a feature flag and is under heavy development, it should not bother
anyone.

Part of story #41084: Retrieve widget configuration

Change-Id: Ia25325cd04542aea33d9ce7a13e2fe48f0f3539e