feat: Add TLP swatch colors support to pie chart
Part of request #44065 pie chart library should know tlp-colors

How to test:
The following charts are still shown:
- Project milestones test results
- Site admin homepage (users, projects, browsers)
- Jenkins widget for Test results (forge a TestResultsPieChartPresenter)
- Project admin > Data > Disk usage

The floating text next to the slices are now 10% darker (as before
commit 413114210ad9f6285c9c9eb8c0ed016ebb34d556). It should offer more
contrast.

If you edit your TestResultsPieChartPresenter and use a TLP swatch CSS
class name for a color, like `tlp-swatch-graffiti-yellow`, the chart
should display that color, and it should use the corresponding Text
color (which is darker) for the floating text next to the slice.

Why?
The class deleted from hudson plugin has been centralized in the lib
itself directly, so that the color dots in the legend always look the
same.

Change-Id: I557f068f82d7946585447e3668663980ff2d8ad9