refactor: Remove popovers from TLP big library
part of request #29982 Split `tlp` in smaller packages

No functional change expected. Popovers in TestManagement, in the "+
New" and the "User avatar" buttons, and the one in the breadcrumbs for
project flags in FlamingParrot and BurningParrot pages
should still function and have styling.
The Popovers page in tlp-doc storybook should still have styling.

Why?
See ADR-0019 for details.

Change-Id: I5f72fd83054d457e63ff01e2b730b3ca258f6801