feat!: Display Text fields in artidoc
Closes story #44020 Artidoc read-only fields: Text fields

Their content is interpreted and displayed the same way as String fields.

They can only be displayed in full row mode. Title and description
semantic are filtered.

Note:
This introduces a Breaking Change in the API: 'string' -> 'text' for
configured field type. As the feature is hidden behind a feature flag
and under development, it should be ok.

*Testing:*

Configure your document and add a text field.

Its value should be displayed

Change-Id: If559f4085292e37e996b5275ad07613b47c2c562