feat: Style date fields in printed PDF
Part of story #44018 Artidoc read-only fields: Date[time] fields

How to test:
- Set up an artidoc with date fields.
- Click on its ellipsis button (…) and export it to a PDF using any
template.
- The date value should not be glued to the label of the field.
- In Firefox, when printing the document as PDF, the date should not be
repeated twice.

Why?
There is an open bug in Firefox [0] where a Custom element that is moved
to another Realm will lose its prototype (so, all its properties and
methods) and revert back to a basic HTMLElement.
This happens to us because print-js will write an <iframe> element with
our special PrinterVersion tree of components, copy them in it, and try
to print that.
To avoid the bug, we define lifecycle callbacks to get back the Custom
element prototype, since it is global to the window.

[0] https://bugzilla.mozilla.org/show_bug.cgi?id=1502814

Change-Id: I723645a5fd38cfd31ec970d7a98f0b0c9addff0d