refactor: WET for service modals init
part of request #44022 pass service modal into vue compat mode

- As project administrator, you can still create, edit and delete
services in Project administration.
- When deleting a service, its name is no longer in bold.
- As site administrator, you can still edit system services in any
Project administration.

Why?
I don't think it's really worth it to pull DOMPurify _just_ to make the
service name in bold. By _not_ doing that, we can reuse tlp-modal's
functions to do exactly what we were doing before, but without the bold
(because we assign the paragraph's textContent).

We also Write Everything Twice to set up the Add and Edit modals because
it's easier to understand. The way we create the app will change in Vue
3, so the tricks with the callbacks might not work. We can simplify if
needed after raising it to Vue 3.

Change-Id: I540fd13707c85199b19cbc7f87929c546023c762