feat: Display PermissionsOnArtifact field
Part of story #44075 Artidoc read-only fields: Permission field

Uses the previous FieldUserGroupsList component to display it as it is a
list of user groups.

Configuration in modal is also done -> you should be able to select this
field.

The PDF render should be good as it uses the same as user groups list
fields.

*Testing:*

Configure your document to display the permissions field. It should be
displayed fine.

Change-Id: I83e6e3494acf45539cd166fc0a605dcfa186f02b