feat: Disable save/cancel buttons while in flight
part of story #44018 Artidoc read-only fields: Date[time] fields

How to test:
- Either set Network conditions in your browser and add a sleep() in the
REST handler to simulate slowness
- When you edit an Artifact section, after saving, the button shows a
spinner icon for all the duration of the GET, PUT, GET sequence.
- Both the "Cancel" and "Save" buttons are disabled for the whole
sequence.

Why?
I renamed the "get-section-in-its-latest-version" file and function
because the real goal is to check for concurrent edition of the
artifact. We never use the section itself and don't care about its
content except to check and raise an error if someone edited the title
or description. I was wondering why we had to do a GET before the PUT,
and the naming did not make it explicit that it was for concurrent
edition.
I also removed some checks for the section types because there are only
two section types, it's not possible to be in another type (at least for
now) so we don't need to check again.

Change-Id: I5612ac91c8765cce7112aa8ca2cffbec9c4fc20f