feat: Backend retrieve StepDefinition values
Part of story #44076 Artidoc read-only fields: Step definition field

For each field value it returns a list of Step with an interpreted
description and expected results. Results can be null if not given for
the step

*Testing:*

Forge your db and add a step def field to your artidoc. In API results
you should see the field values

Change-Id: I811110c31a41b4880303c49c620afe197ef51644