refactor: request #44090 Remove tlp-button-icon-right
No functional change. Icons on the right of buttons should still have
proper spacing from the button text.

Why?
Mixins and Placeholders in the Buttons component are prefixed by a dash
"-" to make them private. They are never reused outside, so should be
internal to the buttons file.

Since buttons are inline-flex, we should use gap instead of margins. It
lets us kill the "-right" class.

Change-Id: If798d33a6fc8dd8ae56c4c212f24f1e2d63f8506