feat: Allow StepsDefinition field in configuration modal
Part of story #44076 Artidoc read-only fields: Step definition field

This field can only be displayed in full row mode.

Note: In the ConfigurationModal, when the field was added you weren't
able to change its display configuration but after save + refresh you
were able to uncheck the box (cannot save but still). It was because
ConfiguredFieldRepresentation only checked for ArtifactLinkField and not
others fields (TextField + StepsDefinition)

*Testing:*

The field should be listed in available ones and you can use it.

Change-Id: I73f69be51536e9fe9c6aafcc1c74962564447d8f