feat: Refresh last update by field
part of story #44018 Artidoc read-only fields: Date[time] fields

How to test:
- Set up an artidoc with the "Last update by" field
- Set up an artifact that was last updated by some other user.
- Edit this artifact in the artidoc. After saving, your user should be
shown in the "Last update by" field, instead of the other user.

Why?
Divs are not allowed as children of spans.
The image was lacking an alt text.
The users_list was not computed, so it did not react to changes in its
props. I also used a one-liner to make sure we always have an array of
users, even for fields with a single user.
I chose to remove the checks on section type for updates. We always get
either a free-text or an artifact section.

Change-Id: I0b1abf80235ef98831b5fde5b2ff18d9638730db