feat: Introduce a new symmetric crypto API
The current legacy one is not yet removed, migration to the new
API will be handled in dedicated contributions.
For now only the parts of Tuleap not persisting values for a long time
have been migrated to it (access keys generation and OAuth2 apps
secrets) since we cannot really migrate them anyway.

The choice has been made to rely on the XChaCha20-Poly1305 construction
and to implement the recommended approach by libsodium to obtain a key
committing AEAD. In an ideal world we would use AEGIS directly but
the libsodium available on EL9 does not support it.

The same encryption key is kept. Ideally we would rotate it but this is
likely to cause additional ops friction that it would be preferable to
not handle at the same time. Having the ability to easily rotate the
encryption key should be handled in a dedicated set of contributions.

Part of request #44081 Internal symmetric cryptographic API should be
resistant against Invisible Salamanders / Confused Deputies attacks

Change-Id: Ib36c92d16ae7b6c1999805b92314af537f9b2cbf