chore: Provide/Inject project_id in Service modal
Part of request #44088 Service modal should be in typescript

Why?
The project_id value is passed directly into the configuration when the app is created.
This value is never modified afterwards in other components.
In such cases, it is best to provide the value when the app is created and inject it in the components that need it.

Change-Id: If6c7c0b5145b7f6e162224b51230fe517ba6e009