chore: Provide/Inject minimal_rank in Service modal
Part of request #44088 Service modal should be in typescript

Why?
The `minimal_rank` value is passed directly into the configuration when the app is created.
This value is never modified afterwards in other components.
Moreover, in certain components, the value was only passed as props to pass it as props to other components.
In such cases, it is best to provide the value when the app is created and inject it in the components that need it.

No functional change expected

Change-Id: Ic49390a01b8228e2010cd4626cf676a4cef7b9eb