chore: Provide/Inject csrf_token in Service modal
Part of request #44088 Service modal should be in typescript

Why?
The `csrf_token` and `csrf_token_name` values were passed directly into the configuration when the app was created.
These values were never modified afterwards in other components.
In such cases, it is best to provide the value when the app is created and inject it in the components that need it.

No functional change expected

Change-Id: I2bce063d33b7639b60905199936e34739ec2ca6a