fix: tracker-creation crashes other scripts in watch mode
Part of request #43330 Convert tracker-creation into vue3

How to test:
- run the following command:
  pnpm --filter=plugin-tracker-creation run watch
- On the tracker creation page, there are no javascript errors on load

Why?
JavascriptAsset creates a script with type=text/javascript, which falls
in the global scope. JavascriptViteAsset creates a script with
type=module, which is correctly isolated (see [0]).
It worked in production because of the minification. Luckily, we did not
run into name collisions. In watch mode, however, some functions like
"addEventListener" were put in the global scope, so they erased previous
functions and wrecked the sidebar, the burningparrot and TLP code.

Also changed the plugin_name in vite config to "trackers", which is the
correct plugin name here. It did not cause issues but would have if we
had used dynamic imports. Also made jQuery external so we are sure it is
fetched from the global scope and not bundled with the app. This reduces
the app size from 332.54 kB to 240.04 kB.

Co-authored-by: Joris Masson <joris.masson@enalean.com>
Co-authored-by: Martin Goyot <martin.goyot@enalean.com>

[0]
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Modules#other_differences_between_modules_and_classic_scripts

Change-Id: Id1729aa36f5b0d3b175fe41157a176d562fa7cc9