refactor: Baseline app runs in Vue 3 compat mode
Implements request #44413 Convert Baseline to Vue 3

How to test:
No functional changes expected in Baseline app:
- You can still add Baselines, display their details, filter by tracker
and delete them.
- You can still create Comparisons, display their details, filter by
tracker and delete them.
- SPA routing should still work: the first breadcrumb leads to the index
page, when you click on links it shows pages.
- Tests should pass.

Why?
I had to pass the project shortname from the backend to be able to build
the correct URL to the base page and give it to the router at init.
Otherwise, the first breadcrumb was broken and lead to nowhere.
I took the opportunity to replace ternaries in ComparisonStatistics by
ngettext calls for plurals.
Change in SCSS was because I saw a warning from autoprefixer during the
build. It should look the same.

Co-authored-by: Joris Masson <joris.masson@enalean.com>
Co-authored-by: Marie Ange Garnier <marie-ange.garnier@enalean.com>
Co-authored-by: Thomas Gorka <thomas.gorka@enalean.com>

Change-Id: I8165640a75fde78fc7a2812e13caf5785701754b