chore: Remove support of Vue 2 apps and libraries
Implements request #44429 End support of Vue 2 apps in build system

No functional change expected.
- CI should be happy
- make scan-vuln-deps should report "No issues found"

Why?
I removed the "Vue 2" configuration for eslint, it is now merged with
the configuration for "Vue 3 apps in compat mode & JavaScript".
I chose to copy the "rule_vue_loader" in all the remaining webpack
config files because now "vue-loader" is installed locally in each app
still using webpack. When we convert them to Vite, we will delete the
webpack config, and that way we will not forget to remove the last one
in "build-system-configurator". It is only 5 lines of config.
All the deleted libs were targetting Vue 2, there are equivalents for
Vue 3.
The TSConfig option "experimentalDecorators" is now disabled, it was
required for Vue Class component, and now they have all been converted.

Change-Id: I6cc870bd6fe9d36a4e3ad238d5e58ad83a245588